update repository to match chosen open src commits

+++ apache-camel/components/camel-box/camel-box-api/src/main/java/org/apache/camel/component/box/api/BoxCommentsManager.java
@@ -26,8 +26,6 @@ import org.apache.camel.RuntimeCamelException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.apache.camel.component.box.api.BoxHelper.buildBoxApiErrorMessage;
-
 /**
  * Provides operations to manage Box comments.
  */
@@ -42,7 +40,7 @@ public class BoxCommentsManager {
 
     /**
      * Create comments manager to manage the comments of Box connection's authenticated user.
-     *
+     * 
      * @param boxConnection - Box connection to authenticated user account.
      */
     public BoxCommentsManager(BoxAPIConnection boxConnection) {
@@ -51,7 +49,7 @@ public class BoxCommentsManager {
 
     /**
      * Add comment to file.
-     *
+     * 
      * @param  fileId  - the id of file.
      * @param  message - the comment's message.
      * @return         The commented file.
@@ -59,28 +57,34 @@ public class BoxCommentsManager {
     public BoxFile addFileComment(String fileId, String message) {
         try {
             LOG.debug("Adding comment to file(id={}) to '{}'", fileId, message);
-            BoxHelper.notNull(fileId, BoxHelper.FILE_ID);
-            BoxHelper.notNull(message, BoxHelper.MESSAGE);
+            if (fileId == null) {
+                throw new IllegalArgumentException("Parameter 'fileId' can not be null");
+            }
+            if (message == null) {
+                throw new IllegalArgumentException("Parameter 'message' can not be null");
+            }
 
             BoxFile fileToCommentOn = new BoxFile(boxConnection, fileId);
             fileToCommentOn.addComment(message);
             return fileToCommentOn;
         } catch (BoxAPIException e) {
             throw new RuntimeCamelException(
-                    buildBoxApiErrorMessage(e), e);
+                    String.format("Box API returned the error code %d%n%n%s", e.getResponseCode(), e.getResponse()), e);
         }
     }
 
     /**
      * Get a list of any comments on this file.
-     *
+     * 
      * @param  fileId - the id of file.
      * @return        The list of comments on this file.
      */
     public List<BoxComment.Info> getFileComments(String fileId) {
         try {
             LOG.debug("Getting comments of file(id={})", fileId);
-            BoxHelper.notNull(fileId, BoxHelper.FILE_ID);
+            if (fileId == null) {
+                throw new IllegalArgumentException("Parameter 'fileId' can not be null");
+            }
 
             BoxFile file = new BoxFile(boxConnection, fileId);
 
@@ -88,33 +92,35 @@ public class BoxCommentsManager {
 
         } catch (BoxAPIException e) {
             throw new RuntimeCamelException(
-                    buildBoxApiErrorMessage(e), e);
+                    String.format("Box API returned the error code %d%n%n%s", e.getResponseCode(), e.getResponse()), e);
         }
     }
 
     /**
      * Get comment information.
-     *
+     * 
      * @param  commentId - the id of comment.
      * @return           The comment information.
      */
     public BoxComment.Info getCommentInfo(String commentId) {
         try {
             LOG.debug("Getting info for comment(id={})", commentId);
-            BoxHelper.notNull(commentId, BoxHelper.COMMENT_ID);
+            if (commentId == null) {
+                throw new IllegalArgumentException("Parameter 'commentId' can not be null");
+            }
 
             BoxComment comment = new BoxComment(boxConnection, commentId);
 
             return comment.getInfo();
         } catch (BoxAPIException e) {
             throw new RuntimeCamelException(
-                    buildBoxApiErrorMessage(e), e);
+                    String.format("Box API returned the error code %d%n%n%s", e.getResponseCode(), e.getResponse()), e);
         }
     }
 
     /**
      * Reply to a comment.
-     *
+     * 
      * @param  commentId - the id of comment to reply to.
      * @param  message   - the message for the reply.
      * @return           The newly created reply comment.
@@ -122,20 +128,23 @@ public class BoxCommentsManager {
     public BoxComment replyToComment(String commentId, String message) {
         try {
             LOG.debug("Replying to comment(id={}) with message={}", commentId, message);
-            BoxHelper.notNull(commentId, BoxHelper.COMMENT_ID);
-            BoxHelper.notNull(message, BoxHelper.MESSAGE);
-
+            if (commentId == null) {
+                throw new IllegalArgumentException("Parameter 'commentId' can not be null");
+            }
+            if (message == null) {
+                throw new IllegalArgumentException("Parameter 'message' can not be null");
+            }
             BoxComment comment = new BoxComment(boxConnection, commentId);
             return comment.reply(message).getResource();
         } catch (BoxAPIException e) {
             throw new RuntimeCamelException(
-                    buildBoxApiErrorMessage(e), e);
+                    String.format("Box API returned the error code %d%n%n%s", e.getResponseCode(), e.getResponse()), e);
         }
     }
 
     /**
      * Change comment message.
-     *
+     * 
      * @param  commentId - the id of comment to change.
      * @param  message   - the new message for the comment.
      * @return           The comment with changed message.
@@ -143,31 +152,36 @@ public class BoxCommentsManager {
     public BoxComment changeCommentMessage(String commentId, String message) {
         try {
             LOG.debug("Changing comment(id={}) message={}", commentId, message);
-            BoxHelper.notNull(commentId, BoxHelper.COMMENT_ID);
-            BoxHelper.notNull(message, BoxHelper.MESSAGE);
-
+            if (commentId == null) {
+                throw new IllegalArgumentException("Parameter 'commentId' can not be null");
+            }
+            if (message == null) {
+                throw new IllegalArgumentException("Parameter 'message' can not be null");
+            }
             BoxComment comment = new BoxComment(boxConnection, commentId);
             return comment.changeMessage(message).getResource();
         } catch (BoxAPIException e) {
             throw new RuntimeCamelException(
-                    buildBoxApiErrorMessage(e), e);
+                    String.format("Box API returned the error code %d%n%n%s", e.getResponseCode(), e.getResponse()), e);
         }
     }
 
     /**
      * Delete comment.
-     *
+     * 
      * @param commentId - the id of comment to delete.
      */
     public void deleteComment(String commentId) {
         try {
             LOG.debug("Deleting comment(id={})", commentId);
-            BoxHelper.notNull(commentId, BoxHelper.COMMENT_ID);
+            if (commentId == null) {
+                throw new IllegalArgumentException("Parameter 'commentId' can not be null");
+            }
             BoxComment comment = new BoxComment(boxConnection, commentId);
             comment.delete();
         } catch (BoxAPIException e) {
             throw new RuntimeCamelException(
-                    buildBoxApiErrorMessage(e), e);
+                    String.format("Box API returned the error code %d%n%n%s", e.getResponseCode(), e.getResponse()), e);
         }
     }
